Issue #14937: Migrated section 5.3 Camel case: defined to follow chapter wise testing
