+++ config/jsoref-spellchecker/whitelist.words
@@ -609,7 +609,6 @@ invalidjavadocposition
 ioe
 ioffset
 IPonr
-IPv
 isfalse
 isindex
 isrutf
